База данных подключена. Все связи корректны, имена колонок и таблиц тоже. Сервер запустился. Был убран функционал в репозиториях, контроллерах и тд. Просто описаны модели.

+++ src/main/java/com/example/school/repository/SchoolEventRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SchoolEventRepository extends JpaRepository<SchoolEvent, Long> {
+public interface SchoolEventRepository extends JpaRepository<SchoolEvent, Integer> {
 }
